Remove non-windows affairs in service.
