+++ tunet-service/Cargo.toml
@@ -9,20 +9,18 @@ license = "MIT"
 tunet-helper = { path = "../tunet-helper" }
 tunet-suggest = { path = "../tunet-suggest" }
 tunet-settings = { path = "../tunet-settings" }
-tokio = { version = "1", features = ["macros", "rt", "signal", "sync"] }
-tokio-stream = { version = "0.1", features = ["sync"] }
-clap = { version = "4", features = ["derive"] }
-cfg-if = "1.0"
-log = "0.4"
-flexi_logger = { version = "0.25", default-features = false, features = [
-    "colors",
-] }
-
-[target.'cfg(target_os = "windows")'.dependencies]
 windows = { version = "0.48", features = [
     "Data_Xml_Dom",
     "Foundation",
     "Networking_Connectivity",
     "UI_Notifications",
 ] }
+tokio = { version = "1", features = ["macros", "rt", "signal", "sync"] }
+tokio-stream = { version = "0.1", features = ["sync"] }
 pin-project = "1.0"
+clap = { version = "4", features = ["derive"] }
+log = "0.4"
+flexi_logger = { version = "0.25", default-features = false, features = [
+    "colors",
+] }
+windows-service = "0.6"
