+++ tunet-service/src/main.rs
@@ -1,9 +1,10 @@
-mod platform_helper;
+mod net_watcher;
+mod notify;
 
 use clap::Parser;
 use flexi_logger::{LogSpecification, Logger};
 use std::{path::PathBuf, pin::pin, sync::Arc};
-use tokio::signal::ctrl_c;
+use tokio::signal::windows::ctrl_c;
 use tokio_stream::StreamExt;
 use tunet_helper::{create_http_client, Result, TUNetConnect, TUNetHelper};
 use tunet_settings::FileSettingsReader;
@@ -28,15 +29,13 @@ async fn main() -> Result<()> {
     let cred = Arc::new(settings.read()?);
     let client = create_http_client()?;
     let c = TUNetConnect::new_with_suggest(None, cred, client).await?;
-    let events = platform_helper::watch()?;
+    let mut ctrlc = ctrl_c()?;
+    let events = net_watcher::watch()?;
     let mut events = pin!(events);
     loop {
         tokio::select! {
-            e = ctrl_c() => {
-                match e {
-                    Ok(()) => log::info!("Ctrl-C received"),
-                    Err(e) => log::error!("{}", e),
-                }
+            _ = ctrlc.recv() => {
+                log::info!("Ctrl-C received");
                 break;
             }
             e = events.next() => {
@@ -57,6 +56,6 @@ async fn login_and_flux(c: &TUNetConnect) -> Result<()> {
     let res = c.login().await?;
     log::info!("{}", res);
     let flux = c.flux().await?;
-    platform_helper::succeeded(flux)?;
+    notify::succeeded(flux)?;
     Ok(())
 }
