+++ tunet-service/src/net_watcher.rs
@@ -5,14 +5,7 @@ use std::{
 };
 use tokio::sync::watch;
 use tokio_stream::{wrappers::WatchStream, Stream};
-use tunet_helper::NetFlux;
-use windows::{
-    core::*,
-    Data::Xml::Dom::XmlDocument,
-    Foundation::EventRegistrationToken,
-    Networking::Connectivity::*,
-    UI::Notifications::{ToastNotification, ToastNotificationManager},
-};
+use windows::{core::*, Foundation::EventRegistrationToken, Networking::Connectivity::*};
 
 pub fn watch() -> Result<impl Stream<Item = ()>> {
     let (tx, rx) = watch::channel(());
@@ -51,28 +44,4 @@ impl Drop for NetworkStatusChangedToken {
         log::debug!("RemoveNetworkStatusChanged");
         NetworkInformation::RemoveNetworkStatusChanged(self.0).unwrap()
     }
-}
-
-pub fn succeeded(flux: NetFlux) -> Result<()> {
-    let xml = format!(
-        r#"<?xml version="1.0" encoding="utf-8"?>
-    <toast>
-        <visual>
-            <binding template="ToastGeneric">
-                <text hint-maxLines="1">登录成功：{0}</text>
-                <text>流量：{1}</text>
-                <text>余额：{2}</text>
-            </binding>
-        </visual>
-    </toast>"#,
-        flux.username, flux.flux, flux.balance
-    );
-    let dom = XmlDocument::new()?;
-    dom.LoadXml(&HSTRING::from(xml))?;
-    let notification = ToastNotification::CreateToastNotification(&dom)?;
-    ToastNotificationManager::CreateToastNotifierWithId(&HSTRING::from(
-        "{1AC14E77-02E7-4E5D-B744-2EB1AE5198B7}\\WindowsPowerShell\\v1.0\\powershell.exe",
-    ))?
-    .Show(&notification)?;
-    Ok(())
 }
