--- tunet-service/src/platform_helper/stub.rs
@@ -1,10 +1,0 @@
-use tokio_stream::{wrappers::WatchStream, Stream};
-use tunet_helper::{anyhow, Result};
-
-pub fn watch() -> Result<impl Stream<Item = ()>> {
-    Err::<WatchStream<()>, _>(anyhow!("不支持的平台"))
-}
-
-pub fn succeeded(_flux: NetFlux) -> Result<()> {
-    Ok(())
-}
